src/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
djangorestframework = "^3.11.1"
django-environ = "^0.4.5"
django-extensions = "^3.0.4"
django-allauth = "^0.42.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
